[TS-286] Rename: StarPage 이름을 StarCardPage로 변경
피그마 페이지 디자인 이름과 동일하게 변경함
