+++ src/App.tsx
@@ -11,7 +11,7 @@ import RestRecordPage from "@/pages/RestRecordPage";
 import OauthSignUp from "@/pages/SignUp/OauthSignUp";
 import SignUp from "@/pages/SignUp/SignUp";
 import StarCardDetailPage from "@/pages/StarCardDetailPage";
-import StarPage from "@/pages/StarPage";
+import StarCardPage from "@/pages/StarCardPage";
 import WithdrawalPage from "@/pages/WithdrawalPage";
 
 const App = () => {
@@ -19,7 +19,7 @@ const App = () => {
 		<BrowserRouter>
 			<Routes>
 				<Route path="/home" element={<HomePage />} />
-				<Route path="/star-card" element={<StarPage />} />
+				<Route path="/star-card" element={<StarCardPage />} />
 				<Route path="/star-card/:id" element={<StarCardDetailPage />} />
 				<Route path="/chart" element={<ChartPage />} />
 				<Route path="/mypage" element={<MyPage />} />
