+++ src/components/StarCard/CategoryTab.tsx
