+++ src/components/StarCard/StarCard.style.ts
